added authorization lv3/4

+++ projekti/app/Http/Controllers/ProjectCollaboratorController.php
@@ -10,6 +10,8 @@ class ProjectCollaboratorController extends Controller
 {
     public function store(Project $project): RedirectResponse
     {
+        $this->authorize('addCollaborator', $project);
+
         if($project->collaborators()->where('user_id', \request()->collaborator_id)->exists())
         {
             return back();
@@ -25,6 +27,8 @@ class ProjectCollaboratorController extends Controller
 
     public function destroy(Project $project, $collaboratorId): RedirectResponse
     {
+        $this->authorize('removeCollaborator', $project);
+
         $project->collaborators()->detach($collaboratorId);
         return redirect(route('project.show', $project));
     }
