+++ projekti/app/Http/Controllers/ProjectTaskController.php
@@ -2,14 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\Task\StoreTaskRequest;
+use App\Http\Requests\Task\UpdateTaskRequest;
 use App\Models\Project;
 use App\Models\Task;
 use Illuminate\Http\Request;
 
 class ProjectTaskController extends Controller
 {
-    public function store(Project $project)
+    public function store(StoreTaskRequest $request, Project $project)
     {
+        $this->authorize('createTask', $project);
+
         $project->tasks()->create(
             [
                 'user_id' => \request()->user()->id,
@@ -20,8 +24,10 @@ class ProjectTaskController extends Controller
         return back();
     }
 
-    public function update(Project $project, Task $task)
+    public function update(UpdateTaskRequest $request, Project $project, Task $task)
     {
+        $this->authorize('updateTask', [$project, $task]);
+
         $task->description = \request()->description;
         $task->update();
         return back();
@@ -29,6 +35,8 @@ class ProjectTaskController extends Controller
 
     public function destroy(Project $project, Task $task)
     {
+        $this->authorize('destroyTask', [$project, $task]);
+
         $task->delete();
         return back();
     }
